scraper/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
ratelimit==2.2.1
pytest==7.4.3
requests-cache==1.1.1
backoff==2.2.1
aiohttp==3.9.1
